Move assertion files
